Bugfix: \setupvimtyping[...][linenumber options]
There was a bug because of which

    \setupvimtyping[...][linenumbering optons]

did not work. This was because these were being set using
\everysetupsyntaxhighlighting which is not called by \setupvimtyping.
So, defined a helper macro which is then called in
\everysetupsyntaxhighlighting and \everysetupvimtyping

It is not possible to pass linenumbering options like so:

    \startvimtyping[...]
      ...
    \stopvimtyping

This is becuase at \startvimtyping be begin a group even before parsing
the options. So, \setuplinenumbering[\currentexternalfilter][..line
numbering options...] gets called inside a group. Since line numbering
works on a page level, these options get ignored.